+++ src/main/java/agaig/justeat/member/annotation/MemberSignInCheck.java
@@ -1,4 +1,4 @@
-package agaig.justeat.annotation;
+package agaig.justeat.member.annotation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
