+++ src/main/java/agaig/justeat/member/domain/Information.java
@@ -1,4 +1,4 @@
-package agaig.justeat.domain;
+package agaig.justeat.member.domain;
 
 import lombok.Builder;
 import lombok.Getter;
