+++ src/main/java/agaig/justeat/member/repository/JdbcTemplateMemberRepository.java
@@ -1,6 +1,6 @@
 //package agaig.justeat.repository;
 //
-//import agaig.justeat.domain.Member;
+//import agaig.justeat.member.domain.Member;
 //import org.springframework.beans.factory.annotation.Autowired;
 //import org.springframework.jdbc.core.JdbcTemplate;
 //import org.springframework.jdbc.core.RowMapper;
