+++ src/main/java/agaig/justeat/member/repository/MemberRepository.java
@@ -1,6 +1,6 @@
-package agaig.justeat.repository;
+package agaig.justeat.member.repository;
 
-import agaig.justeat.domain.Member;
+import agaig.justeat.member.domain.Member;
 import org.apache.ibatis.annotations.Mapper;
 
 import java.util.List;
