+++ src/main/resources/application.properties
@@ -7,4 +7,4 @@ spring.datasource.password=12345678
 spring.jpa.open-in-view=true
 server.port = 3070
 mybatis.mapper-locations=classpath:mybatis/mapper/*.xml
-mybatis.type-aliases-package=agaig.justeat.domain
+mybatis.type-aliases-package=agaig.justeat.member.domain
