+++ src/main/resources/mybatis/mapper/InformationMapper.xml
@@ -3,7 +3,7 @@
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="agaig.justeat.repository.InformationRepository">
+<mapper namespace="agaig.justeat.member.repository.InformationRepository">
     <select id="select" parameterType="long" resultType="Information">
         SELECT info_id, title, content, writer, view_cnt, comment_cnt, reg_date
         FROM information
